requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
Pillow>=9.0.0
asyncio
aiosqlite>=0.19.0
sqlmodel>=0.0.16